Renames a bunch of files
Adds three new directories, common, v16 and 17.
