+++ rpc/src/main/scala/org/bitcoins/rpc/client/common/MempoolRpc.scala
@@ -1,8 +1,7 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.common
 
 import org.bitcoins.core.crypto.DoubleSha256Digest
 import org.bitcoins.rpc.jsonmodels.{ GetMemPoolEntryResult, GetMemPoolInfoResult, GetMemPoolResult }
-import org.bitcoins.rpc.serializers.BitcoindJsonReaders._
 import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 import play.api.libs.json.{ JsBoolean, JsString }
