+++ rpc/src/main/scala/org/bitcoins/rpc/client/common/P2PRpc.scala
@@ -1,12 +1,11 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.common
 
 import java.net.URI
 
 import org.bitcoins.core.protocol.blockchain.Block
 import org.bitcoins.rpc.jsonmodels._
-import org.bitcoins.rpc.serializers.BitcoindJsonReaders._
-import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 import play.api.libs.json.{ JsBoolean, JsNumber, JsString }
+import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 
 import scala.concurrent.Future
