+++ rpc/src/main/scala/org/bitcoins/rpc/client/common/RawTransactionRpc.scala
@@ -1,13 +1,11 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.common
 
 import org.bitcoins.core.crypto.{ DoubleSha256Digest, ECPrivateKey }
 import org.bitcoins.core.currency.Bitcoins
 import org.bitcoins.core.protocol.BitcoinAddress
+import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 import org.bitcoins.core.protocol.transaction.{ Transaction, TransactionInput }
 import org.bitcoins.rpc.jsonmodels.{ FundRawTransactionResult, GetRawTransactionResult, RpcTransaction, SignRawTransactionResult }
-import org.bitcoins.rpc.serializers.BitcoindJsonReaders._
-import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
-import org.bitcoins.rpc.serializers.BitcoindJsonWriters._
 import play.api.libs.json._
 
 import scala.concurrent.Future
