+++ rpc/src/main/scala/org/bitcoins/rpc/client/common/WalletRpc.scala
@@ -1,14 +1,13 @@
-package org.bitcoins.rpc.client
+package org.bitcoins.rpc.client.common
 
 import org.bitcoins.core.crypto.{ DoubleSha256Digest, ECPrivateKey, ECPublicKey }
 import org.bitcoins.core.currency.Bitcoins
 import org.bitcoins.core.protocol.BitcoinAddress
 import org.bitcoins.core.protocol.blockchain.MerkleBlock
 import org.bitcoins.core.protocol.transaction.Transaction
-import org.bitcoins.rpc.client.RpcOpts.AddressType
+import org.bitcoins.rpc.client.common.RpcOpts.AddressType
 import org.bitcoins.rpc.jsonmodels._
 import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
-import org.bitcoins.rpc.serializers.BitcoindJsonWriters._
 import play.api.libs.json._
 
 import scala.concurrent.Future
